Update hand_index_querier to build simpler catchment data parquet
hand_index_querier updated to just create a table with refs to relevant
data. This reflect the modification made to the hand index to just pass
hydrotable csv refs when queried instead of hydrotable data to avoid
duplication and reduce hand index load step complexity and runtime.

Also updated the restart nomad script and moved it to the local nomad
directory