Adjust master_metrics.csv columns and add tags to inundate stage
Added bench_src and scenario to inundate stage. This is useful info for
the inundate jobs since they are inundated with flow information for
specific scenarios.

Edited some of the metadata columns in make_master_metrics.py to have
the same header names as the OWP version of master_metrics.csv and also
added a couple hard coded columns for extent configuration and the json
path column